backend/requirements.txt
fastapi
uvicorn
sqlalchemy
xgboost>=2.0.3
lightgbm>=4.3.0
catboost>=1.2.2
pandas
numpy>=1.26.4
scikit-learn>=1.3.2
matplotlib
binance
python-binance
python-dotenv
python-multipart
cryptography
requests
aiohttp
httpx
alembic
psycopg2-binary
psutil
apscheduler==3.10.4
prometheus-client==0.21.0
torch>=2.0.0

# Architecture v2 Dependencies
redis>=5.0.0
structlog>=23.0.0
asyncpg>=0.29.0

# RL v3 Dependencies (PPO-based reinforcement learning)
gymnasium>=1.0.0
stable-baselines3>=2.0.0
ccxt>=4.0.0

# Auth and security dependencies
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4

# Additional utilities
PyYAML>=6.0.0
aiosqlite>=0.19.0

# NOTE: SQLAlchemy-Utils was previously pinned here for investigation of PyUp
# advisory 42194. A scan of the codebase showed no direct usage of
# SQLAlchemy-Utils APIs (EncryptedType etc.), so we remove the pin and allow
# the project to avoid installing it unless explicitly required by developers.
# If dev tooling needs it, add it to backend/requirements-dev.txt instead.
